feat(eur-settings): add initial implementation of backend settings
