+++ crates/app/eur-settings/Cargo.toml
@@ -16,3 +16,4 @@ dirs = { workspace = true }
 specta = { version = "=2.0.0-rc.22", features = ["derive", "function"] }
 eur-fs = { workspace = true }
 notify = { version = "8.2.0" }
+eur-prompt-kit = { workspace = true }
