+++ crates/app/eur-settings/src/settings/general.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, Type)]
+#[serde(rename_all = "camelCase")]
+pub struct GeneralSettings {
+    /// Whether to start the app on autostart
+    pub autostart: bool,
+}
