+++ crates/app/eur-settings/src/settings/launcher.rs
@@ -1,0 +1,10 @@
+use crate::hotkey::Hotkey;
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, Type, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct LauncherSettings {
+    #[serde(default)]
+    pub hotkey: Hotkey,
+}
