+++ crates/app/eur-settings/src/settings/mod.rs
@@ -1,0 +1,11 @@
+mod backend;
+mod general;
+mod hover;
+mod launcher;
+mod telemetry;
+
+pub use backend::*;
+pub use general::*;
+pub use hover::*;
+pub use launcher::*;
+pub use telemetry::*;
